+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.2.1
+- Temp files are no longer re-written if they are already up-to-date in both verify and build mode 
+- New flag `--needed/-N` and corresponding mode `InMemoryBuild` that stores the fresh output in memory and only writes the file if different
+
 ## 0.2.0
 - Verify mode no longer re-writes temp files if they are already up-to-date.